SCB-508 Add cache size metric
